requirements.txt
pymodbus
python-dotenv